Sequence id generation strategy has been provided. Product entity bug fixed. GET method of fetching all the data has been tested.

+++ src/main/java/com/app/repository/EnteryStockPlaceRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EnteryStockPlaceRepository extends JpaRepository<EnteryStockPlace, Integer> {
-}
+public interface EnteryStockPlaceRepository extends JpaRepository<EnteryStockPlace, Long> {}
