+++ src/main/java/com/app/repository/OutputStockPlaceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OutputStockPlaceRepository extends JpaRepository<OutputStockPlace, Integer> {
+public interface OutputStockPlaceRepository extends JpaRepository<OutputStockPlace, Long> {
 }
