+++ src/main/java/com/app/repository/TestProductsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TestProductsRepository extends JpaRepository<TestProducts, Integer> {
+public interface TestProductsRepository extends JpaRepository<TestProducts, Long> {
 }
